+++ CHANGELOG.md
@@ -12,6 +12,7 @@ CentOS-6 6.8 x86_64 - MySQL 5.1.
 - Adds a change log (`CHANGELOG.md`).
 - Adds support for semantic version numbered tags.
 - Adds minor code style changes to the Makefile for readability.
+- Adds support for running `shpec` functional tests with `make test`.
 
 ### 1.7.2 - 2016-10-02